1) Disease Image Model/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
albumentations>=1.3.0
scikit-learn>=1.0.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.0.0
streamlit